requirements.txt
absl-py==0.2.0
alabaster==0.7.10
astor==0.6.2
Babel==2.6.0
backcall==0.1.0
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
Cython==0.28.3
decorator==4.3.0
dev==0.4.0
docutils==0.14
easydict==1.7
entrypoints==0.2.3
gast==0.2.0
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imagesize==1.0.0
imgaug==0.4
ipykernel==4.8.2
ipyparallel==6.2.1
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.1.6
kiwisolver==1.0.1
lightgbm==2.1.1
Markdown==2.6.11
MarkupSafe==1.0
mask-rcnn==2.1
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nose==1.3.7
notebook==5.5.0
numpy==1.18.5
opencv-python==3.4.1.15
packaging==17.1
pandocfilters==1.4.2
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==8.4.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.17.2
scikit-learn==0.19.1
scipy==1.5.2
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.1.0
tensorboard==1.15
tensorflow-gpu==1.15
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.1
